Input mistake proofing the code.
Added a Sensors subsystem to call sensors from different commands. Automated more off of button presses. Split the controls to two Xbox Controllers. Added an second option on the gamepad to deposit at the Amp rather than speaker. Added an automatic retract when the Intake lowers to intake a Note.